Move TypeMappingInfo to Npgsql.Internal.TypeHandling namespace
