+++ src/Npgsql/Internal/TypeHandling/TypeHandlerResolverFactory.cs
@@ -1,7 +1,4 @@
 using System;
-using System.Data;
-using Npgsql.TypeMapping;
-using NpgsqlTypes;
 
 namespace Npgsql.Internal.TypeHandling
 {
@@ -22,22 +19,4 @@ namespace Npgsql.Internal.TypeHandling
         internal static TypeMappingInfo? GetMappingByValueDependentValue(this TypeHandlerResolverFactory factory, object value)
             => factory.GetDataTypeNameByValueDependentValue(value) is { } dataTypeName ? factory.GetMappingByDataTypeName(dataTypeName) : null;
     }
-}
-
-public class TypeMappingInfo
-{
-    public TypeMappingInfo(NpgsqlDbType? npgsqlDbType, string? dataTypeName, Type clrType)
-        => (NpgsqlDbType, DataTypeName, ClrTypes) = (npgsqlDbType, dataTypeName, new[] { clrType });
-
-    public TypeMappingInfo(NpgsqlDbType? npgsqlDbType, string? dataTypeName, params Type[] clrTypes)
-        => (NpgsqlDbType, DataTypeName, ClrTypes) = (npgsqlDbType, dataTypeName, clrTypes);
-
-    public NpgsqlDbType? NpgsqlDbType { get; }
-    DbType? dbType;
-    public DbType DbType
-        => dbType ??= NpgsqlDbType is null ? DbType.Object : GlobalTypeMapper.NpgsqlDbTypeToDbType(NpgsqlDbType.Value);
-    public string? DataTypeName { get; }
-    public Type[] ClrTypes { get; }
-
-    internal void Reset() => dbType = null;
 }
